requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.1.0
click==8.1.3
coverage==7.2.1
Django==4.1.7
djangorestframework==3.14.0
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==17.5.0
iniconfig==2.0.0
isort==5.12.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pluggy==1.0.0
psycopg2==2.9.5
pytest==7.2.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
